Add separate file for delay handling
